format fix for float()

+++ src/gui/page_optimizer/notebook_optimizer/optimizer_panels/swarm_settings_panels/panel_pso_basic.py
@@ -105,9 +105,9 @@ class PSO_Basic_Panel(wx.Panel):
         noError = True
         try:
             self.numParticles = self.fieldnumParticles.GetValue()
-            self.weights = [np.float(self.fieldWeights1.GetValue()),
-                            np.float(self.fieldWeights2.GetValue()), 
-                            np.float(self.fieldWeights3.GetValue())]
+            self.weights = [self.fieldWeights1.GetValue(),
+                            self.fieldWeights2.GetValue(), 
+                            self.fieldWeights3.GetValue()]
             self.vlim = self.fieldVlim.GetValue()
             self.toleranceVal = self.fieldTolerance.GetValue()
             self.maxIter = self.fieldMaxIter.GetValue()
